feat[manifests]: Allow BYO argocd
Restructuring the manifests to allow consumers of AVP to bring their own argocd
base by using the manifests under `manifests/plugin`.

This allows consumers to integrate more easily with other CMPs or alternate
versions of argo that the AVP maintainers haven't patched the manifests to
include.

Because the resources they patch are now not included in the kustomization bases
in manifests/plugins, the standard `patchesStrategicMerge` strategy will fail.
To fix this we switch to the `patches` interface and provide an explicit target
for each patch.